requirements.txt
langchain
langchain-community
langchain_groq
langchain_huggingface
chromadb
python-dotenv
streamlit
sentence-transformers
pandas
numpy